+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added a new option on the character list to remove a character from the addon.
 - Added a scrollbar to the character list in case you're an extreme altoholic (60+ characters).
 
 ## v1.4.0 - 2026-03-11